requirements.txt
pytest~=7.4.4
Appium-Python-Client~=2.0.0
selenium~=4.0.0